Split out the chunk function to it's own file with tests
