+++ report/data_sources/hubspot/client.py
@@ -3,11 +3,12 @@ import json
 import os.path
 from dataclasses import dataclass
 from enum import Enum
-from itertools import islice
 from typing import Callable, Generator, Iterable, List, Optional, Set
 
 from hubspot import HubSpot
 from hubspot.crm.associations import BatchInputPublicObjectId
+
+from report.data_sources.chunk import chunk
 
 
 @dataclass
@@ -21,21 +22,6 @@ class Field:
     def __post_init__(self):
         if not self.key:
             self.key = self.hs_field
-
-
-def chunk(items: Iterable, chunk_size: int):
-    """Return items in multiple chunks of a maximum size.
-
-    :param items: Items to break into chunks
-    :param chunk_size: The maximum size of a chunk (last may be shorter)
-    """
-
-    # Create an iterable we can consume from
-    iter_items = iter(items)
-    # Use the two-argument form of iter to specify when to stop. This will
-    # cause iter to call the lambda until it returns an empty tuple, which
-    # should consume the iterable above.
-    return iter(lambda: tuple(islice(iter_items, chunk_size)), ())
 
 
 @dataclass
